chore(dependencies): update nitro version and adjust server start script
This commit modifies the `package.json` to update the nitro dependency version and changes the server start script to point to the new output directory. Additionally, the preview script's host has been updated to use `127.0.0.1` instead of `0.0.0.0`. These changes aim to improve project configuration and ensure compatibility with the latest setup.